Multithread handling of incoming AIS packets
Handling AIS packets takes time, as an AIS custom message is extracted
from the packet and is added to a buffer that records the source of the
message. Over time, this can lead to overflow on the AisBus side as
packets are not consumed as fast as they are being received.

This introduces a `received but yet unhandled` buffer on the receiver
side and uses a fixed thread pool to process and remove packets stored
in that buffer. A configuration option is also introduced to restrain
this buffer in size and allow overflowing.